# Improved AVX2 adler32 performance