@KungFuJesus 2022-01-09
The before:
```
2022-01-09T11:56:10-05:00
Running ./adler32_benchmark
Run on (28 X 4100.04 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x14)
  L1 Instruction 32 KiB (x14)
  L2 Unified 1024 KiB (x14)
  L3 Unified 19712 KiB (x1)
Load Average: 0.20, 0.25, 0.17
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
adler32_c_bench               42731 ns        42730 ns        72151
adler32_ssse3_bench           12177 ns        12177 ns       100000
adler32_avx2_bench             9635 ns         9634 ns       100000
adler32_avx512_bench           6598 ns         6597 ns       100000
adler32_avx512vnni_bench       5415 ns         5415 ns       100000
```

The after:

```
2022-01-09T11:56:56-05:00
Running ./adler32_benchmark
Run on (28 X 4099.54 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x14)
  L1 Instruction 32 KiB (x14)
  L2 Unified 1024 KiB (x14)
  L3 Unified 19712 KiB (x1)
Load Average: 0.13, 0.22, 0.16
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
adler32_c_bench               43102 ns        43101 ns        73120
adler32_ssse3_bench           12219 ns        12218 ns       100000
adler32_avx2_bench             7295 ns         7295 ns       100000
adler32_avx512_bench           6419 ns         6419 ns       100000
adler32_avx512vnni_bench       5391 ns         5391 ns       100000
```