@KungFuJesus 2022-01-13
Using @nmoinvaz's improved benchmarks from his PR, here's the improvements with better measurements, as taken from a low clock speed haswell -U series laptop (Thinkpad x240):

Before: 
```
adler32/avx2/2048            185 ns          203 ns      3869309
adler32/avx2/4096            346 ns          371 ns      2123199
adler32/avx2/32768          2393 ns         2576 ns       272217
adler32/avx2/262144        19845 ns        21539 ns        32886
adler32/avx2/2097152      164407 ns       179753 ns         3852
adler32/avx2/4194304      382851 ns       410988 ns         1703
```

After:
```
adler32/avx2/2048            111 ns          122 ns      5733796
adler32/avx2/4096            209 ns          231 ns      2918230
adler32/avx2/32768          1666 ns         1795 ns       397546
adler32/avx2/262144        17293 ns        18668 ns        29235
adler32/avx2/2097152      141763 ns       152206 ns         4095
adler32/avx2/4194304      342989 ns       368295 ns         1865
```